requirements.txt
asgiref==3.7.2
coverage==5.5
deprecation==2.1.0
Django==4.2.6
django-admin==2.0.2
django-cors-headers==4.3.0
django-excel-response2==3.0.6
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
eyed3==0.9.7
filetype==1.2.0
isoweek==1.3.3
packaging==23.2
Pillow==10.1.0
pydub==0.25.1
python-dateutil==2.8.2
pytz==2023.3.post1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
toml==0.10.2
tzlocal==5.2
xlwt==1.3.0
